Ea-2502 added the logic for multiple aggregation/proxy, clean up for EDMManifest class, fixed jnuit.
