+++ src/main/java/eu/europeana/iiif/service/ManifestService.java
@@ -275,7 +275,7 @@ public class ManifestService {
     public ManifestV2 generateManifestV2 (String json, boolean addFullText, URL fullTextApi)    {
         long start = System.currentTimeMillis();
         Object document = com.jayway.jsonpath.Configuration.defaultConfiguration().jsonProvider().parse(json);
-        ManifestV2 result = EdmManifestMapping.getManifestV2(document);
+        ManifestV2 result = EdmManifestMappingV2.getManifestV2(document);
 
         if (addFullText) {
             try {
@@ -304,7 +304,7 @@ public class ManifestService {
     public ManifestV3 generateManifestV3 (String json, boolean addFullText, URL fullTextApi)  {
         long start = System.currentTimeMillis();
         Object document = com.jayway.jsonpath.Configuration.defaultConfiguration().jsonProvider().parse(json);
-        ManifestV3 result = EdmManifestMapping.getManifestV3(document);
+        ManifestV3 result = EdmManifestMappingV3.getManifestV3(document);
 
         if (addFullText) {
             try {
